pyproject.toml
[tool.poetry]
name = "pocketsmith-api-spec"
version = "0.1.0"
description = ""
authors = [
    'Zach "theY4Kman" Kanzler <they4kman@gmail.com>',
    'PocketSmith Ltd <api@pocketsmith.com>',
]


[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.dev-dependencies]
"ruamel.yaml" = "^0.16.10"
pytest = "^6.2.5"
pytest-camel-collect = "^1.0.2"
pytest-icdiff = "^0.5"


[tool.pytest.ini_options]
addopts = "--verbosity=2 --tb=short"

# Only find tests underneath the tests/ directory
testpaths = ["tests"]

# Only search for tests within files matching these patterns
python_files = "tests.py test_*.py"

# Discover tests within classes matching these patterns
# NOTE: the dash represents a word boundary (functionality provided by pytest-camel-collect)
python_classes = "Test-* Describe-* Context-* With-* Without-* For-* When-* If-* Case-*"

# Only methods matching these patterns are considered tests
python_functions = "test_* it_* its_*"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
